Fix wrong -t command; Fix usage/help
This commit fixes the wrong -t command (test command) and combines the usage and
help command into one switch case since it's using the same output anyway.

Also adds `basename $0` to the usage so the name of the called file is inserted
instead of `vhost` (in case someone re-names it on their machine).