[v0.0.2] Notifications, Tests, Removal
This release contains a new test function that will check if folders and files
can be accessed by the script.

There's also a new set of functions called `remove` and `removeHost`. `remove`
will prompt the user to if they really want to remove the given host and
`removeHost` will then delete it.

Added a Changelog to keep track on changes; Updated Readme with latest
information and warning about tested OS (only Ubuntu).